Delay in last universe fixed
While framing it was discovered that the last universe would be delayed or display on its own. This was very noticeable in any strobe effects.  By recording by video we slowed down the video and could easily see the problem.  After making loop counter corrections and advancing the loop count after the LED function the framing was consistant and we were able to get the last universe to frame the same as the rest.